Exercise 3.1
Starting with the Vehicle class from the exercises in Chapter 2, create
an extended class called PassengerVehicle to add a capability for
counting the number of seats available in the car and the number
currently occupied. Provide a new main method in PassengerVehicle to
create a few of these objects and print them out.